디렉터리 구조 변경
page 디렉터리 안에 컴포넌트별로 디렉터리를 나눔
