+++ src/page/case/CaseAdvices.tsx
